+++ cmd/drone-server/inject_plugin.go
@@ -20,6 +20,7 @@ import (
 	"github.com/drone/drone/plugin/admission"
 	"github.com/drone/drone/plugin/config"
 	"github.com/drone/drone/plugin/converter"
+	"github.com/drone/drone/plugin/converter/starlark"
 	"github.com/drone/drone/plugin/registry"
 	"github.com/drone/drone/plugin/secret"
 	"github.com/drone/drone/plugin/validator"
@@ -79,7 +80,9 @@ func provideConfigPlugin(client *scm.Client, contents core.FileService, conf spe
 func provideConvertPlugin(client *scm.Client, conf spec.Config) core.ConvertService {
 	return converter.Combine(
 		converter.Legacy(false),
-		converter.Starlark(false),
+		starlark.New(
+			conf.Starlark.Enabled,
+		),
 		converter.Jsonnet(
 			conf.Jsonnet.Enabled,
 		),