+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- support for starlark scripts in core.
+- support for executing pipelines in debug mode.
+
 ## [1.9.2]
 ### Added
 - update go-scm dependency to fix